+++ sklearn/metrics/classification.py
@@ -16,6 +16,7 @@ the lower the better
 #          Joel Nothman <joel.nothman@gmail.com>
 #          Noel Dawe <noel@dawe.me>
 #          Jatin Shah <jatindshah@gmail.com>
+#          Saurabh Jha <saurabh.jhaa@gmail.com>
 # License: BSD 3 clause
 
 from __future__ import division
@@ -1376,14 +1377,20 @@ def log_loss(y_true, y_pred, eps=1e-15, normalize=True, sample_weight=None):
     return _weighted_sum(loss, sample_weight, normalize)
 
 
-def hinge_loss(y_true, pred_decision, pos_label=None, neg_label=None):
+def hinge_loss(y_true, pred_decision, labels=None):
     """Average hinge loss (non-regularized)
 
-    Assuming labels in y_true are encoded with +1 and -1, when a prediction
-    mistake is made, ``margin = y_true * pred_decision`` is always negative
-    (since the signs disagree), implying ``1 - margin`` is always greater than
-    1.  The cumulated hinge loss is therefore an upper bound of the number of
-    mistakes made by the classifier.
+    In binary class case, assuming labels in y_true are encoded with +1 and -1,
+    when a prediction mistake is made, ``margin = y_true * pred_decision`` is
+    always negative (since the signs disagree), implying ``1 - margin`` is
+    always greater than 1.  The cumulated hinge loss is therefore an upper
+    bound of the number of mistakes made by the classifier.
+
+    In multiclass case, the function expects that either all the labels are
+    included in y_true or an optional labels argument is provided which
+    contains all the labels. The multilabel margin is calculated according
+    to Crammer-Singer's method. As in the binary case, the cumulated hinge loss
+    is an upper bound of the number of mistakes made by the classifier.
 
     Parameters
     ----------
@@ -1394,6 +1401,9 @@ def hinge_loss(y_true, pred_decision, pos_label=None, neg_label=None):
     pred_decision : array, shape = [n_samples] or [n_samples, n_classes]
         Predicted decisions, as output by decision_function (floats).
 
+    labels : array, optional, default None
+        Contains all the labels for the problem. Used in multiclass hinge loss.
+
     Returns
     -------
     loss : float
@@ -1402,6 +1412,16 @@ def hinge_loss(y_true, pred_decision, pos_label=None, neg_label=None):
     ----------
     .. [1] `Wikipedia entry on the Hinge loss
             <http://en.wikipedia.org/wiki/Hinge_loss>`_
+
+    .. [2] Koby Crammer, Yoram Singer. On the Algorithmic
+           Implementation of Multiclass Kernel-based Vector
+           Machines. Journal of Machine Learning Research 2,
+           (2001), 265-292
+
+    .. [3] 'L1 AND L2 Regularization for Multiclass Hinge Loss Models
+           by Robert C. Moore, John DeNero.
+           <http://www.ttic.edu/sigml/symposium2011/papers/
+           Moore+DeNero_Regularization.pdf>'
 
     Examples
     --------
@@ -1420,27 +1440,56 @@ def hinge_loss(y_true, pred_decision, pos_label=None, neg_label=None):
     >>> hinge_loss([-1, 1, 1], pred_decision)  # doctest: +ELLIPSIS
     0.30...
 
+    In the multiclass case:
+    >>> X = np.array([[0], [1], [2], [3]])
+    >>> Y = np.array([0, 1, 2, 3])
+    >>> labels = np.array([0, 1, 2, 3])
+    >>> est = svm.LinearSVC()
+    >>> est.fit(X, Y)
+    LinearSVC(C=1.0, class_weight=None, dual=True, fit_intercept=True,
+         intercept_scaling=1, loss='l2', max_iter=1000, multi_class='ovr',
+         penalty='l2', random_state=None, tol=0.0001, verbose=0)
+    >>> pred_decision = est.decision_function([[-1], [2], [3]])
+    >>> y_true = [0, 2, 3]
+    >>> hinge_loss(y_true, pred_decision, labels)  #doctest: +ELLIPSIS
+    0.56...
     """
-    # TODO: multi-class hinge-loss
     check_consistent_length(y_true, pred_decision)
+    pred_decision = check_array(pred_decision, ensure_2d=False)
     y_true = column_or_1d(y_true)
-    pred_decision = column_or_1d(pred_decision)
+    y_true_unique = np.unique(y_true)
+    if y_true_unique.size > 2:
+        if (labels is None and pred_decision.ndim > 1 and
+                (np.size(y_true_unique) != pred_decision.shape[1])):
+            raise ValueError("Please include all labels in y_true "
+                             "or pass labels as third argument")
+        if labels is None:
+            labels = y_true_unique
+        le = LabelEncoder()
+        le.fit(labels)
+        y_true = le.transform(y_true)
+        mask = np.ones_like(pred_decision, dtype=bool)
+        mask[np.arange(y_true.shape[0]), y_true] = False
+        margin = pred_decision[~mask]
+        margin -= np.max(pred_decision[mask].reshape(y_true.shape[0], -1),
+                         axis=1)
 
-    # the rest of the code assumes that positive and negative labels
-    # are encoded as +1 and -1 respectively
-    lbin = LabelBinarizer(neg_label=-1)
-    y_true = lbin.fit_transform(y_true)[:, 0]
+    else:
+        # Handles binary class case
+        # this code assumes that positive and negative labels
+        # are encoded as +1 and -1 respectively
+        pred_decision = column_or_1d(pred_decision)
+        pred_decision = np.ravel(pred_decision)
 
-    if len(lbin.classes_) > 2 or (pred_decision.ndim == 2
-                                  and pred_decision.shape[1] != 1):
-        raise ValueError("Multi-class hinge loss not supported")
-    pred_decision = np.ravel(pred_decision)
+        lbin = LabelBinarizer(neg_label=-1)
+        y_true = lbin.fit_transform(y_true)[:, 0]
 
-    try:
-        margin = y_true * pred_decision
-    except TypeError:
-        raise TypeError("pred_decision should be an array of floats.")
+        try:
+            margin = y_true * pred_decision
+        except TypeError:
+            raise TypeError("pred_decision should be an array of floats.")
+
     losses = 1 - margin
-    # The hinge doesn't penalize good enough predictions.
+    # The hinge_loss doesn't penalize good enough predictions.
     losses[losses <= 0] = 0
     return np.mean(losses)